Move tests out of trait
ParserTests was a trait containing tests that were shared between the Regex & Runtime parsers.

Now we just have the Runtime parser, we can move all the tests into the TemplateTest.
